Moved SymbolPickerButtonV to its own file
